requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anytree==2.13.0
attrs==25.3.0
beautifulsoup4==4.13.4
build==1.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
docopt==0.6.2
fastjsonschema==2.21.2
fonttools==4.58.4
frozenlist==1.7.0
idna==3.10
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.6.4
nbformat==5.10.4
numpy==2.3.0
packaging==25.0
pandas==2.3.0
patsy==1.0.1
pigar==2.1.7
pillow==11.2.1
pip-chill==1.0.3
pip-tools==7.5.0
pipgrip==0.11.0
pipreqs==0.4.13
platformdirs==4.4.0
propcache==0.3.2
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
statsmodels==0.14.4
threadpoolctl==3.6.0
traitlets==5.14.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.14
wheel==0.45.1
yarg==0.1.10
yarl==1.20.1
